Notion/backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-filter==2.4.0
nltk
psycopg2-binary==2.9.1
python-dotenv==0.19.1
torch
spacy==3.1.2
scikit-learn==0.24.2
transformers==4.9.2
numpy==1.21.2
speechrecognition==3.8.1
Pillow==8.3.1
pytesseract==0.3.8


celery==5.1.2
django-celery-beat==2.2.0
django-celery-results==2.1.0


redis==3.5.3

